requirements.txt
APScheduler ==3.9.1
kafka-python==2.0.2
elasticsearch==8.2.0